backend/agents/requirements.txt
langchain>=0.1.0
anthropic==0.18.1
python-dotenv==1.0.0
mistralai==0.0.12
typing>=3.7.4
groq==0.4.2
openai==1.12.0
llama-index-core>=0.10.0
fastapi>=0.110.0
torch>=2.2.0
pydantic>=2.6.0
aiohttp>=3.9.0
PyMuPDF>=1.23.0
arxiv>=2.1.0
tenacity>=8.2.0
youtube-transcript-api>=0.6.2
requests>=2.31.0
newspaper3k>=0.2.8
langchain-core>=0.1.27
langchain-anthropic==0.1.4
asyncio>=3.4.3
duckduckgo-search==6.3.7
primp==0.8.1